requirements.txt
# requirements.txt
google-generativeai
python-dotenv
Pillow
flask
flask-sqlalchemy
flask-login
werkzeug